pyproject.toml
[tool.poetry]
name = "org-mgmt"
version = "0.1.0"
description = ""
authors = ["0xajinkya <kahaneajinkya051@gmail.com>"]
readme = "README.md"
package-mode = false
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

# Core dependencies
streamlit = "1.47.0"
pandas = "2.3.1"
requests = "2.32.4"
pydantic = "2.11.7"
python-dotenv = "1.1.1"
sqlacodegen = "3.0.0"
sqlalchemy = "2.0.41"
psycopg2-binary = "2.9.10"

# UI Enhancements
streamlit-shadcn-ui = "0.1.18"
streamlit-extras = "0.7.5"
streamlit-avatar = "0.1.3"
streamlit-camera-input-live = "0.2.0"
streamlit-card = "1.0.2"
streamlit-embedcode = "0.1.2"
streamlit-image-coordinates = "0.3.1"
streamlit-keyup = "0.3.0"
streamlit-toggle-switch = "1.0.2"
streamlit-vertical-slider = "2.5.5"
streamlit_faker = "0.0.4"
st-theme = "1.2.3"
st-annotated-text = "4.0.2"

# Visualization
altair = "5.5.0"
matplotlib = "3.10.3"
plotly = "6.2.0"
pydeck = "0.9.1"

# Utilities
protobuf = ">=3.20,<6"
cloudpickle = ">=2.2.1,<=3.0.0"
annotated-types = "0.7.0"
beautifulsoup4 = "4.13.4"
black = "25.1.0"
blinker = "1.9.0"
cachetools = "6.1.0"
certifi = "2025.7.14"
charset-normalizer = "3.4.2"
click = "8.2.1"
contourpy = "1.3.2"
cryptography = "45.0.5"
cycler = "0.12.1"
faker = "37.4.2"
favicon = "0.7.0"
fonttools = "4.59.0"
gitpython = "3.1.45"
greenlet = "3.2.3"
htbuilder = "0.9.0"
idna = "3.10"
inflect = "7.5.0"
jinja2 = "3.1.6"
jsonschema = "4.25.0"
jsonschema-specifications = "2025.4.1"
kiwisolver = "1.4.8"
lxml = "6.0.0"
markdown = "3.8.2"
markdownlit = "0.0.7"
markupsafe = "3.0.2"
more-itertools = "10.7.0"
mypy-extensions = "1.1.0"
narwhals = "1.48.1"
numpy = "2.3.2"
packaging = "25.0"
pathspec = "0.12.1"
pillow = "11.3.0"
platformdirs = "4.3.8"
prometheus-client = "0.22.1"
pyarrow = "18.1.0"
pycparser = "2.22"
pydantic-core = "2.33.2"
pyjwt = "2.10.1"
pymdown-extensions = "10.16"
pyopenssl = "25.1.0"
pyparsing = "3.2.3"
python-dateutil = "2.9.0.post0"
pytz = "2025.2"
pyyaml = "6.0.2"
referencing = "0.36.2"
rpds-py = "0.26.0"
six = "1.17.0"
smmap = "5.0.2"
snowflake-connector-python = "3.16.0"
snowflake-snowpark-python = "1.35.0"
sortedcontainers = "2.4.0"
soupsieve = "2.7"
tenacity = "9.1.2"
toml = "0.10.2"
tomlkit = "0.13.3"
tornado = "6.5.1"
typeguard = "4.4.4"
typing-inspection = "0.4.1"
typing-extensions = "4.14.1"
tzdata = "2025.2"
tzlocal = "5.3.1"
urllib3 = "2.5.0"
validators = "0.35.0"
watchdog = "6.0.0"
wheel = "0.45.1"
boto3 = "1.39.13"
botocore = "1.39.13"
entrypoints = "0.4"
jmespath = "1.0.1"
s3transfer = "0.13.1"
cffi = "1.17.1"
asn1crypto = "1.5.1"
attrs = "25.3.0"
flake8 = "^7.3.0"
mypy = "^1.17.1"
isort = "^6.0.1"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
